Add some sound effects, particles and fixes
Added sound effects for the door, campfire and BGM for the intro scene(forest sounds) and for the dungeon(some creepy music), will add more sounds in the future and adjust the existing ones.

Now only the playerController script is being disabled during cutscenes instead of the player itself(making it so that the timeline dosen't need a Player tag for the scene to be loaded).

Added particles effects that act like fog in the Intro Scene and the Traps Room Scene (will have to adjust them as they are causing the game to lag quite a lot).

Fixed a few more lighting issues.

Started adding some more decorations at the dungeon entrance and will add more around the Intro Scene.

Now using trees with LOD's in the Intro Scene.

*Reminder to change the lights in the Dungeon Entrance Part Scene.