Scripts fixes, arrow traps, changed NPC's
Changed the door script, now using animations for opening and closing (need to fix the audio).
Added arrow traps in the Dungeon Part 3(not working as intended, will have to fix the arrow speed for better timings).
To enter the Dungeon Part 3 Scene you now need a "skull key" that opens the door, upon picking up the "Red Skull" you will activate the trigger that closes the door when leaving the scene.
Changed the NPC's with low poly ones instead of the ones from Mixamo (the animations for them are still from Mixamo).
Organized the scripts a little better.