Add NPC, door to part2, fireballs, respawn points and more.
Added an NPC in the Dungeon Entrance Part that the player can interact with to start a dialogue(not skippable) which explains what they have to do.

Added a key for the Traps Room door.

The spheres in the Traps Room must now be activated in order to open a door that leads to the Dungeon Part 2, activating all 4 spheres will trigger a cutscene which shows the door opening.

The spheres are now bouncing using animations instead of doing that through scripts.

Added particles to torches.

Added respawn points in the Traps Room.

Added a key in the labyrinth for the Labyrinth Exit Door.

Added fireballs in the labyrinth that respawn the player if they collide with it.

Added sound to the fireballs (might have to redo that).

Modified the teleporter to the Mage Room Part, the player will now have to press a key and wait for some seconds before teleporting.

Added particles to the teleporters.

A few more minor fixes.