Modify most of the cutscene scripts, add NPC AI and more
Most of the cutscene/dialogue scripts are now using Events, which lowers the amount of code needed.
Added an NPC in the small basement cell in the Dungeon Part 2 which the player has to rescue after doing the mage missions, the mage will give the player a key that unlocks the door to the cell.
Added dialogue with the NPC in the cell and 2 missions after the end of the dialogue. (The missions are not implemented yet).
The NPC in the cell has animations for idle sitting, standing up, idle standing, walking, he follows the player around after accepting a mission.
Added a Keybinds panel in the options which shows the controls.
Tried to add a loading screen. (still has some bugs)

I finally know how animations work. (more or less)