Add final mission
Added the final mission which takes the player to the main gate which leads out of the dungeon. (for now the gate will just open leading to no where, will add an end screen soon)
Added animations for the main gate.
After completing the mage missions, he will now place a key on the table which opens the door to the small cell.

After talking to the NPC in the small cell, you can pick one of the two missions (only 1 can be done, but both will lead to the same spot).
The first mission takes the player back to the mage which will tell them to leave through the main gate.
The second mission takes the player straight to the main gate.

Added missing particles to a torch in the hallway to the mage room.
Added all doors in a DOORS game object parent for easy access.